+++ services/web/server/src/simcore_service_webserver/folders/_models.py
@@ -18,10 +18,9 @@ from models_library.utils.common_validators import (
     null_or_none_str_to_none_validator,
 )
 from models_library.workspaces import WorkspaceID
-from pydantic import BeforeValidator, ConfigDict, Field
-from servicelib.request_keys import RQT_USERID_KEY
+from pydantic import BaseModel, BeforeValidator, ConfigDict, Field
 
-from .._constants import RQ_PRODUCT_KEY
+from .._constants import RQ_PRODUCT_KEY, RQT_USERID_KEY
 
 _logger = logging.getLogger(__name__)
 
@@ -88,3 +87,12 @@ class FolderSearchQueryParams(
 
 class FolderTrashQueryParams(RemoveQueryParams):
     ...
+
+
+class _FolderWorkspacesPathParams(BaseModel):
+    folder_id: FolderID
+    workspace_id: Annotated[
+        WorkspaceID | None, BeforeValidator(null_or_none_str_to_none_validator)
+    ] = Field(default=None)
+
+    model_config = ConfigDict(extra="forbid")
