+++ services/web/server/src/simcore_service_webserver/folders/_workspaces_handlers.py
@@ -1,61 +1,20 @@
-import functools
 import logging
-from typing import Annotated
 
 from aiohttp import web
-from models_library.folders import FolderID
-from models_library.utils.common_validators import null_or_none_str_to_none_validator
-from models_library.workspaces import WorkspaceID
-from pydantic import BaseModel, BeforeValidator, ConfigDict, Field
 from servicelib.aiohttp import status
 from servicelib.aiohttp.requests_validation import parse_request_path_parameters_as
-from servicelib.aiohttp.typing_extension import Handler
 
 from .._meta import api_version_prefix as VTAG
-from ..folders.errors import FolderAccessForbiddenError, FolderNotFoundError
 from ..login.decorators import login_required
-from ..projects.exceptions import ProjectInvalidRightsError, ProjectNotFoundError
 from ..security.decorators import permission_required
-from ..workspaces.errors import WorkspaceAccessForbiddenError, WorkspaceNotFoundError
 from . import _workspaces_api
-from ._models import FoldersRequestContext
+from ._exceptions_handlers import handle_plugin_requests_exceptions
+from ._models import FoldersRequestContext, _FolderWorkspacesPathParams
 
 _logger = logging.getLogger(__name__)
 
 
-def _handle_folders_workspaces_exceptions(handler: Handler):
-    @functools.wraps(handler)
-    async def wrapper(request: web.Request) -> web.StreamResponse:
-        try:
-            return await handler(request)
-
-        except (
-            ProjectInvalidRightsError,
-            FolderNotFoundError,
-            WorkspaceNotFoundError,
-        ) as exc:
-            raise web.HTTPNotFound(reason=f"{exc}") from exc
-
-        except (
-            ProjectNotFoundError,
-            FolderAccessForbiddenError,
-            WorkspaceAccessForbiddenError,
-        ) as exc:
-            raise web.HTTPForbidden(reason=f"{exc}") from exc
-
-    return wrapper
-
-
 routes = web.RouteTableDef()
-
-
-class _FolderWorkspacesPathParams(BaseModel):
-    folder_id: FolderID
-    workspace_id: Annotated[
-        WorkspaceID | None, BeforeValidator(null_or_none_str_to_none_validator)
-    ] = Field(default=None)
-
-    model_config = ConfigDict(extra="forbid")
 
 
 @routes.put(
@@ -64,7 +23,7 @@ class _FolderWorkspacesPathParams(BaseModel):
 )
 @login_required
 @permission_required("folder.update")
-@_handle_folders_workspaces_exceptions
+@handle_plugin_requests_exceptions
 async def replace_project_workspace(request: web.Request):
     req_ctx = FoldersRequestContext.model_validate(request)
     path_params = parse_request_path_parameters_as(_FolderWorkspacesPathParams, request)
